Refactor stealth.py
